+++ client/src/components/form/Login/index.tsx
@@ -1,48 +1,13 @@
 import React from 'react';
-import styled from 'styled-components';
-import { postLogin } from '@src/apis';
-import { INPUT, TOAST_MESSAGE, PLACEHOLDER_TXT } from '@utils/constant';
-import { FORM } from '@utils/styleConstant';
-import Form from './common/Form';
-import useInput from '@hooks/useInput';
+import { Container, Input, Button, ModalToggleSpan } from './style';
+import Form from '@src/components/form/Form';
 import { useUserFns } from '@contexts/userContext';
 import { useToast } from '@src/hooks/useToast';
+import useInput from '@hooks/useInput';
+import { postLogin } from '@src/apis';
+import { INPUT, TOAST_MESSAGE, PLACEHOLDER_TXT } from '@utils/constant';
 import { isEmail, isPassword } from '@utils/utils';
-
-const Container = styled.div`
-  display: flex;
-  flex-direction: column;
-  width: ${FORM.loginWidth}rem;
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  position: absolute;
-  top: 20%;
-`;
-
-const Input = styled.input`
-  font-size: ${({ theme }) => theme.fontSizes.lg};
-`;
-
-const Button = styled.button`
-  ${({ theme }) => theme.flexCenter}
-  background-color: ${({ theme }) => theme.colors.green};
-  padding: ${({ theme }) => theme.paddings.sm};
-  color: ${({ theme }) => theme.colors.white};
-  border-radius: 1rem;
-`;
-
-const ModalToggleSpan = styled.span`
-  width: 100%;
-  margin-top: ${({ theme }) => theme.margins.lg};
-  color: ${({ theme }) => theme.colors.blue};
-  text-align: center;
-  cursor: pointer;
-  :hover {
-    color: ${({ theme }) => theme.colors.primary};
-  }
-`;
+import { FORM } from '@utils/styleConstant';
 
 interface LoginProps {
   onClickModalToggle: React.MouseEventHandler<HTMLButtonElement>;
